+++ src/pl/plajer/villagedefense3/commands/CommandData.java
@@ -1,0 +1,37 @@
+/*
+ * Village Defense 3 - Protect villagers from hordes of zombies
+ * Copyright (C) 2018  Plajer
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package pl.plajer.villagedefense3.commands;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+/**
+ * @author Plajer
+ * <p>
+ * Created at 03.05.2018
+ */
+@Data
+@AllArgsConstructor
+public class CommandData {
+
+    private String text;
+    private String command;
+    private String description;
+
+}