+++ src/pl/plajer/villagedefense3/items/SpecialItemManager.java
@@ -27,9 +27,7 @@ import java.util.HashMap;
  */
 public class SpecialItemManager {
 
-
     private static HashMap<String, SpecialItem> specialItems = new HashMap<>();
-
 
     public static void addEntityItem(String name, SpecialItem entityItem) {
         specialItems.put(name, entityItem);